+++ source/voxelman/platform/cursoricon.d
@@ -1,0 +1,17 @@
+/**
+Copyright: Copyright (c) 2017 Andrey Penechko.
+License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
+Authors: Andrey Penechko.
+*/
+
+module voxelman.platform.cursoricon;
+
+enum CursorIcon
+{
+	arrow,
+	ibeam,
+	crosshair,
+	hand,
+	hresize,
+	vresize,
+}